requirements.txt
fastmcp>=2.8.1
requests>=2.32.4
python-dotenv